// src/enums/commonEnums.ts
// Enums for Table Type
export enum eResultCodes
    {
        R_SUCCESS = 0,
        R_DB_ERROR = 1,
        R_SERVICE_GET_DATA_ERROR = 2,
        R_SERVICE_POST_DATA_ERROR = 3,
        R_NO_DATA_FOUND = 4,
        R_AUTHENTICATION_FAILED = 5,
        R_UNAUTHORIZED = 6,
        R_UNKNOWN = 7,
        R_INVALID_LOGIN_ID = 8,
        R_INVALID_PASSWORD = 9,
        R_SERVICE_ERROR = 10,
        R_SERVICE_UNAVAILABLE = 11,
        R_INVALID_REQUEST = 12,
        R_INVALID_NUMBER = 13,
        R_NOT_FOUND = 14,
        R_NETWORK_ERROR_OR_SERVER_ERROR = 15,
        R_CREATED = 16,
        R_INVALID_SSO_PASSWORD = 17,
        R_NOT_AVAILABLE = 18,
        R_INTERNAL_SERVER_ERROR = 19,
        R_UNUSED = 20,
        R_MULTIPLE_RECORDS = 21,
        R_BAD_REQUEST = 22,
        R_USER_NOT_ACTIVE = 23,
        R_USER_VALIDITY_EXPIRED = 24,
        R_DUPLICATE = 25,
        R_FUNCTIONAL_ERROR = 26,
        R_MAIL_ERROR = 27,
        R_TENENT_NOT_FOUND = 28,
        R_OUT_OF_SCHEDULE = 30,
        R_DEVICE_IN_USE = 31,
        R_ALREADY_LOGGED_IN = 32,
        R_NOT_LOGGED_IN = 33,
        R_NO_ACCESS_ON_INSTALLATION = 34,
        R_ASSET_INFO_NOT_FOUND = 35,
        R_VALUE_BEYOND_LIMIT = 36,
        R_RECORD_NOT_EXIST = 37,
        R_CONFIGURATION_MISSING = 38,
        R_CONFLICT_OCCURRED = 39,
        R_USER_HAS_ACTIVE_INSTALLATION = 40,
        R_METHOD_NOT_ALLOWED = 41,
        R_FIELD_BLANK = 42,
        R_USER_NOT_FOUND = 43,
        R_REQUIRED = 44
    }